mllm_requirements.txt
accelerate==1.12.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
av==15.1.0
beautifulsoup4==4.14.3
bitsandbytes==0.41.3
black==26.1.0
certifi==2026.1.4
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
DataProperty==1.1.0
datasets==4.5.0
decord==0.6.0
dill==0.4.0
distlib==0.4.0
distro==1.9.0
easydict==1.13
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
evaluate==0.4.6
exceptiongroup==1.3.1
fastapi==0.129.0
ffmpy==1.0.0
filelock==3.24.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==0.17.3
httpx==0.24.0
huggingface_hub==0.36.2
identify==2.6.16
idna==3.11
importlib_resources==6.5.2
isort==7.0.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonlines==4.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latex2mathml==3.78.1
latex2sympy2==1.9.1
-e git+https://github.com/haotian-liu/LLaVA@c121f0432da27facab705978f83c4ada465e46fd#egg=llava

loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
math-verify==0.1.0
matplotlib==3.10.8
mbstrdecoder==1.1.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.4.2
nltk==3.9.2
nodeenv==1.10.0
numexpr==2.14.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.21.0
openai-clip==1.0.1
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
orjson==3.11.7
packaging==25.0
pandas==2.3.3
pathspec==1.0.4
pathvalidate==3.3.1
peft==0.18.1
pillow==10.4.0
platformdirs==4.9.2
portalocker==3.2.0
pre_commit==4.5.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pybind11==3.0.2
pycocoevalcap==1.2
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
ruff==0.15.1
sacrebleu==2.6.0
safetensors==0.7.0
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.2.3
sentencepiece==0.1.99
sentry-sdk==2.53.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
sqlitedict==2.1.0
starlette==0.52.1
svgwrite==1.4.3
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==0.6.13
tokenizers==0.22.2
tomli==2.4.0
tomlkit==0.12.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
tqdm-multiprocess==0.0.11
transformers==4.57.6
transformers-stream-generator==0.0.5
triton==3.6.0
typepy==1.3.4
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==20.37.0
wandb==0.25.0
wavedrom==2.0.3.post3
wcwidth==0.6.0
websockets==11.0.3
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2026.2.4
zss==1.2.0
zstandard==0.25.0
